+++ src/fallbackRest.ts
@@ -20,6 +20,7 @@ import * as serializer from 'proto3-json-serializer';
 import {defaultToObjectOptions} from './fallback';
 import {FetchParameters} from './fallbackServiceStub';
 import {hasTextDecoder, hasTextEncoder, isNodeJS} from './featureDetection';
+import {GoogleError} from './googleError';
 import {transcode} from './transcoding';
 
 if (!hasTextEncoder() || !hasTextDecoder()) {
@@ -91,10 +92,7 @@ export function decodeResponse(
   const decodedString = new TextDecoder().decode(response);
   const json = JSON.parse(decodedString);
   if (!ok) {
-    const error = Object.assign(
-      new Error(json['error']['message']),
-      json.error
-    );
+    const error = GoogleError.parseHttpError(json);
     throw error;
   }
   const message = serializer.fromProto3JSON(rpc.resolvedResponseType!, json);